testing

+++ stabilizer/src/main/java/com/hazelcast/stabilizer/tests/map/helpers/KeyInc.java
@@ -5,7 +5,7 @@ import java.io.Serializable;
 public class KeyInc implements Serializable {
 
     public int key=0;
-    public double inc=0;
+    public int inc=0;
 
     public KeyInc() {
     }
